Move a bunch of application code to the application directory
