+++ autogpt/config/__init__.py
@@ -3,10 +3,12 @@ This module contains the configuration classes for AutoGPT.
 """
 from .ai_config import AIConfig
 from .config import Config, ConfigBuilder, check_openai_api_key
+from .prompt_config import PromptConfig
 
 __all__ = [
     "check_openai_api_key",
     "AIConfig",
     "Config",
     "ConfigBuilder",
+    "PromptConfig",
 ]
