+++ autogpt/config/ai_config.py
@@ -5,16 +5,9 @@ A module that contains the AIConfig class object that contains the configuration
 from __future__ import annotations
 
 import os
-import platform
 from pathlib import Path
-from typing import TYPE_CHECKING, Optional
 
-import distro
 import yaml
-
-if TYPE_CHECKING:
-    from autogpt.models.command_registry import CommandRegistry
-    from autogpt.prompts.generator import PromptGenerator
 
 # Soon this will go in a folder where it remembers more stuff about the run(s)
 SAVE_FILE = str(Path(os.getcwd()) / "ai_settings.yaml")
@@ -53,8 +46,8 @@ class AIConfig:
         self.ai_role = ai_role
         self.ai_goals = ai_goals
         self.api_budget = api_budget
-        self.prompt_generator: PromptGenerator | None = None
-        self.command_registry: CommandRegistry | None = None
+        # HACK: Existing code wants this to be here. Can't fix quite yet.
+        self.prompt_generator = None
 
     @staticmethod
     def load(ai_settings_file: str = SAVE_FILE) -> "AIConfig":
@@ -85,7 +78,6 @@ class AIConfig:
             for goal in config_params.get("ai_goals", [])
         ]
         api_budget = config_params.get("api_budget", 0.0)
-        # type: Type[AIConfig]
         return AIConfig(ai_name, ai_role, ai_goals, api_budget)
 
     def save(self, ai_settings_file: str = SAVE_FILE) -> None:
@@ -108,58 +100,3 @@ class AIConfig:
         }
         with open(ai_settings_file, "w", encoding="utf-8") as file:
             yaml.dump(config, file, allow_unicode=True)
-
-    def construct_full_prompt(
-        self, config, prompt_generator: Optional[PromptGenerator] = None
-    ) -> str:
-        """
-        Returns a prompt to the user with the class information in an organized fashion.
-
-        Parameters:
-            None
-
-        Returns:
-            full_prompt (str): A string containing the initial prompt for the user
-              including the ai_name, ai_role, ai_goals, and api_budget.
-        """
-
-        prompt_start = (
-            "Your decisions must always be made independently without"
-            " seeking user assistance. Play to your strengths as an LLM and pursue"
-            " simple strategies with no legal complications."
-            ""
-        )
-
-        from autogpt.prompts.prompt import build_default_prompt_generator
-
-        if prompt_generator is None:
-            prompt_generator = build_default_prompt_generator(config)
-        prompt_generator.goals = self.ai_goals
-        prompt_generator.name = self.ai_name
-        prompt_generator.role = self.ai_role
-        prompt_generator.command_registry = self.command_registry
-        for plugin in config.plugins:
-            if not plugin.can_handle_post_prompt():
-                continue
-            prompt_generator = plugin.post_prompt(prompt_generator)
-
-        if config.execute_local_commands:
-            # add OS info to prompt
-            os_name = platform.system()
-            os_info = (
-                platform.platform(terse=True)
-                if os_name != "Linux"
-                else distro.name(pretty=True)
-            )
-
-            prompt_start += f"\nThe OS you are running on is: {os_info}"
-
-        # Construct full prompt
-        full_prompt = f"You are {prompt_generator.name}, {prompt_generator.role}\n{prompt_start}\n\nGOALS:\n\n"
-        for i, goal in enumerate(self.ai_goals):
-            full_prompt += f"{i+1}. {goal}\n"
-        if self.api_budget > 0.0:
-            full_prompt += f"\nIt takes money to let you run. Your API budget is ${self.api_budget:.3f}"
-        self.prompt_generator = prompt_generator
-        full_prompt += f"\n\n{prompt_generator.generate_prompt_string(config)}"
-        return full_prompt
